[WIP] Back to the old prefix behavior in `company-jupyter-repl`
This needs to be changed. `company-grab-symbol` will only grab 'bar' when
cursor is at | in the following:

    foo.bar|

We would like it to grab the whole 'foo.bar' and let the kernel figure out what
to do. Similarly for other operators like '->'. The problem that needs to be
fixed is to properly replace just the 'bar' part if 'foo.bar' is the prefix. I
believe the `match` command of a company backend will be of use.