Add `jupyter-region-context`
Also update the default code context for completion requests to be the current
line up to `point`.

This allows completion in python kernels to work as expected. Previously when
sending the whole line, the python completer would not return expected results.
For example, if you had something like

os.mkdir(os.path.|, 0777)

then you would not receive any completions for the os.path namespace, only for
the global namespace.