jupyter-code-context: Return the full code block in an `org-mode` context
Better support for completions in the Jupyter org-mode source blocks

A line context is often insufficient. For example, the Julia kernel will parse
up to the current point and return dictionary keys as completions or method
arguments. Similarly a Python kernel may use Jedi for completion which does
static analysis of the code block to resolve import paths for completion
results.