jupyter-with-control-code-handling: Properly handle \r
If a \r appears as the last character in the inserted text of BODY,
`jupyter-handle-control-codes` does not delete the whole line. Instead it makes
the character invisible. It waits until future output comes in to delete the
whole line. The issue was that `jupyter-with-control-code-handling` was not
taking into account the \r character from previous output.