Remove `jupyter-ansi-color-apply-on-region`
The only modification compared to `ansi-color-apply-on-region` was that it did
not call `delete-and-extract-region` on the region containing ANSI escapes,
instead the ANSI escaped were rendered invisible. This can be done by `let`
binding `delete-and-extract-region` to a custom function while calling
`ansi-color-apply-on-region` instead.

The downside is that we are now depending on an implementation detail of
`ansi-color-apply-on-region`.