jupyter-return-delayed-thunk: Use `thunk-delay`
This makes it so that when a delayed value is bound, via
`jupyter-mlet*`, it always results in the same value as was computed
the first time it was bound.  Previously the computation associated
with the delayed value would occur each time it was bound, resulting
in a newly computed value each time.