Implement interrupt_request messages
close #495

* jupyter-base.el (jupyter-session-endpoints): Consider control_port
in the connection info of a session.
(jupyter-channel-from-request-type): New function.

* jupyter-client.el (jupyter-interrupt-kernel): Send an
interrupt_request when the kernel supports it.
(jupyter-handle-interrupt-reply): New client handler.

* jupyter-kernel-process.el (jupyter-kernel): Don't try to create a
launch-able kernel object when the connection info is provided.
(jupyter-zmq-io): Consider the control channel.
(jupyter-interrupt): Raise an error if trying to interrupt a kernel
when the mode of interruption is message based.

* jupyter-monads.el (jupyter-sent): Determine the channel to send on
based off of the request type.
(jupyter-request): Use `jupyter-channel-from-request-type` to check if
the channel of a request is stdin.

* jupyter-repl.el (jupyter-connect-repl): Spoof a valid kernelspec so
that `jupyter-interrupt-kernel` can be used.

* jupyter-zmq-channel-ioloop.el (initialize-instance): Consider the
control channel.
(jupyter-zmq-channel-ioloop--recv-messages): Update documentation.