Support source block reference resolution
So that Jupyter blocks can be used as named references in contexts
such as noweb, calls to `org-sbe`, and as the value of :var header
arguments.

The current implementation, a tentative one, returns a string
containing the concatenation of all the text/plain representations of
the results of the source block wrapped in a call to
`org-babel-script-escape` to convert simple values to elisp
representations.

fixes #222
fixes #525

* jupyter-base.el (jupyter-mime-value): New function.

* jupyter-org-client.el (jupyter-org-execute-async-p): New function.
(jupyter-generate-request): Use it.
(jupyter-handle-execute-request): Move inline-block result insertion.
(jupyter-org--do-insert-result): To here. Account for changes in
`jupyter-org--add-result`.
(jupyter-handle-display-data)
(jupyter-org-pending-async-results): Account for changes in
`jupyter-org--add-result`.
(jupyter-org--add-result): Change the arguments to accept mime bundles
as well as Org elements.
(jupyter-org--coalesce-stream-results): Don't convert stream results
to scalar elements.
(jupyter-org-sync-results): Return an interesting result when
":results silent" or ":results none" is specified so that it can be
interpreted by callers of `org-execute-src-block`.

* ob-jupyter.el (org-babel-jupyter--execute): Account for changes in
the format of `jupyter-org-request-results`.

* test/jupyter-test.el (org-babel-src-block-name-resolution): New test.