Generalize `jupyter-read-expression`
It was mainly for reading an expression to evaluate, now it is up to
the caller what the purpose of it is for.

* jupyter-client.el (jupyter-eval-expression-history): Rename to
`jupyter-expression-history`.
(jupyter--read-expression-history): Rename to
`jupyter--expression-history`.
(jupyter--setup-minibuffer): New function.
(jupyter-read-expression): Add an optional prompt argument, use
`jupyter--setup-minibuffer`.  Update to take into account for changes
in variable names.
(jupyter-kernel-language): Make the client argument optional.
(jupyter-eval-string-command)
(jupyter-eval): Provide prompt to `jupyter-read-expression`.

* jupyter-org-client.el (jupyter-org--define-key-filter): Take into
account change to `jupyter-kernel-language`.

* jupyter-repl.el (jupyter-read-expression): Remove.