Don't attempt to handle messages if the Org buffer disappears
This is a partial solution to #531 which would still make the code
more robust to the fact that the buffer can indeed be killed while an
asynchronous block is running.  A more complete solution would be to
stop processing messages from kernels whose Org buffers have been
deleted or, in the case of a deletion it may still be useful to
partially process the output, e.g. redirect stream messages to an
external buffer and display results in other ways outside of the Org
buffer.

* jupyter-org-client.el (jupyter-org-with-point-at): New macro.  Same
as `org-with-point-at`, but does nothing if the request's marker
points nowhere.
(jupyter-org--goto-error-string)
(jupyter-handle-error)
(jupyter-handle-execute-reply, jupyter-org-result)
(jupyter-org--clear-async-indicator)
(jupyter-org-inserted-result): Replace calls to `org-with-point-at` to
`jupyter-org-with-point-at`.