Cache notebook server response in one instance
When accessing directories using TRAMP the function calls end up
hitting `jupyter-api-server-accessible-p` indirectly from the calls
`jupyter-tramp-with-api-connection` ->
`jupyter-tramp-server-from-file-name`.  Since
`jupyter-tramp-with-api-connection` is used throughout all of the
TRAMP file operations it can be called a lot which means a server
request just to check if the server is accessible before making the
request for the TRAMP operation since
`jupyter-api-server-accessible-p` did not cache its result.  This
change, caches the result so that TRAMP can perform all of its file
operations while only having to do an initial check of the server
availability.

* jupyter-rest-api (jupyter-api--response-cache)
(jupyter-api--response-cahce-expiry): New variables.
(jupyter-api-server-accessible-p): Cache response value.