Light refactoring and suppress an error in one instance
* jupyter-org-client.el (jupyter-org--with-src-block-client): New
function. same as what the non-private version did, in addition, use
the new NOERROR argument of `org-babel-jupyter-session-initiated-p`.
(jupyter-org-with-src-block-client): Move the work to the above so as
to reduce clutter of the macro expansion.
(jupyter-org--insert-stream): Remove a `setq`.

* ob-jupyter.el (org-babel-jupyter-session-key): Add optional
argument, NOERROR, that returns nil if a session key could not be
formed instead of raising an error.
(org-babel-jupyter-session-initiated-p): Add optional argument, same
as above.

* jupyter-repl.el (jupyter-repl-interaction-mode-line): Use `pcase`.