Refactor source block caching
* jupyter-org-client.el (jupyter-org--src-block-cache): Allow a new
form for the cache, to indicate that it is invalid but still keep the
information around for potential use as information of the most
recently visited source block.
(jupyter-org--src-block-beg, jupyter-org--src-block-end): Remove.
(jupyter-org--same-src-block-p): Rename to
`jupyter-org--at-cached-src-block-p`.
(jupyter-org--set-current-src-block): Rename to
`jupyter-org--set-src-block-cache`.  Handle invalidation semantics.
(jupyter-org-with-src-block-bounds): New macro.
(jupyter-org-src-block-params): New function.
(jupyter-org-when-in-src-block)
(jupyter-org--with-src-block-client): Update to use new/modified
functions.
(jupyter-org-with-src-block-client): Rework documentation.
(jupyter-code-context): Remove use of removed functions.