Be safer when handling types during completion
* jupyter-client.el (jupyter-completion-construct-candidates): Don't
allow a potential call to `max` with zero arguments.  Consider the
potential that there may not be types for all matches. In this case,
the type information displayed will be wrong for some matches and not
appear for others.  In the type information there is available the
actual match that it is supposed to be associated with which could
be used to assign the type information to the exact match.