Don't use a "system internal mounted directory" for delete tests.
This is a GLib-specific bug, so perhaps the "fix" should be more
narrowly scoped to those systems.  At any rate, when Jupyter notebook
tries to delete a file or directory under /tmp, it throws this error
from the python send2trash library:

    gi.repository.GLib.GError: g-io-error-quark: Trashing on system
    internal mounts is not supported (15)

which ultimately manifests as a permission denied error in these tests.

Fixes emacs-jupyter/jupyter#620